+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add Data Synchronisation Detail Component to notify administrative users about cas / dogu behaviour (#119)
+
 ### Changed
 - Update ces-theme-tailwind to v0.3.8
+
 
 ## [v1.13.0-1] - 2024-02-29
 ### Added