+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.14.0-1] - 2024-03-11
 ### Added
 - Add email notification support for newly created users via csv import (#125)
 - Add Data Synchronisation Detail Component to notify administrative users about cas / dogu behaviour (#119)